+++ test/function/cpp/entry.cpp
@@ -16,12 +16,66 @@
  * limitations under the License.
  */
 
+#include <cstdlib>
 #include <gtest/gtest.h>
 #include <stumpless.hpp>
 
 using namespace stumpless;
 
 namespace {
+
+  class CppEntryTest : public::testing::Test {
+  protected:
+    Entry *basic_entry;
+    const char *basic_app_name = "stumpless-cpp-testing";
+    const char *basic_msgid = "basic-msgid";
+    const char *basic_message = "This is a basic entry.";
+
+    virtual void
+    SetUp( void ) {
+      basic_entry = new Entry( Facility::USER,
+                               Severity::INFO,
+                               basic_app_name,
+                               basic_msgid,
+                               "This is a basic entry." );
+    }
+
+    virtual void
+    TearDown( void ) {
+      delete basic_entry;
+    }
+
+  };
+
+  TEST_F( CppEntryTest, GetAppName ) {
+    const char *result;
+
+    result = basic_entry->GetAppName(  );
+    EXPECT_STREQ( result, basic_app_name );
+
+    free( ( void * ) result );
+  }
+
+  TEST_F( CppEntryTest, GetMessage ) {
+    const char *result;
+
+    result = basic_entry->GetMessage(  );
+    EXPECT_STREQ( result, basic_message );
+
+    free( ( void * ) result );
+  }
+
+  TEST_F( CppEntryTest, GetMsgid ) {
+    const char *result;
+
+    result = basic_entry->GetMsgid(  );
+    EXPECT_STREQ( result, basic_msgid );
+
+    free( ( void * ) result );
+  }
+
+  /* non-fixture tests */
+
   TEST( BuildEntry, WithElement ) {
     Entry my_entry( Facility::USER,
                     Severity::INFO,